backend/pyproject.toml
[project]
name = "fundpal-backend"
version = "0.1.0"
description = "Backend for FundPal Financial Assistant"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "database", "routes"]
